moved all rolls to the items
