+++ src/actors/context-menus/rune-magic-context-menu.ts
@@ -9,15 +9,11 @@ import {
   RqgError,
 } from "../../system/util";
 import { ItemTypeEnum } from "../../data-model/item-data/itemTypes";
-import { RuneMagicChatHandler } from "../../chat/runeMagicChatHandler";
 import { ContextMenuRunes } from "./contextMenuRunes";
 import { RqgItem } from "../../items/rqgItem";
 import { Rqid } from "../../system/api/rqidApi";
 
-export const runeMagicMenuOptions = (
-  actor: RqgActor,
-  token: TokenDocument | undefined
-): ContextMenu.Item[] => [
+export const runeMagicMenuOptions = (actor: RqgActor): ContextMenu.Item[] => [
   {
     name: localize("RQG.Game.RollChat"),
     icon: ContextMenuRunes.RollViaChat,
@@ -36,16 +32,7 @@ export const runeMagicMenuOptions = (
       const itemId = getRequiredDomDataset(el, "item-id");
       const item = actor.items.get(itemId);
       assertItemType(item?.data.type, ItemTypeEnum.RuneMagic);
-      await RuneMagicChatHandler.roll(
-        item,
-        item?.data.data.points,
-        0,
-        0,
-        0,
-        0,
-        actor,
-        ChatMessage.getSpeaker({ actor: actor, token: token })
-      );
+      item.abilityRoll();
     },
   },
   {
