+++ src/actors/context-menus/spirit-magic-context-menu.ts
@@ -1,5 +1,4 @@
 import { RqgActorSheet } from "../rqgActorSheet";
-import { SpiritMagicChatHandler } from "../../chat/spiritMagicChatHandler";
 import { RqgActor } from "../rqgActor";
 import {
   assertItemType,
@@ -15,10 +14,7 @@ import { ContextMenuRunes } from "./contextMenuRunes";
 import { RqgItem } from "../../items/rqgItem";
 import { Rqid } from "../../system/api/rqidApi";
 
-export const spiritMagicMenuOptions = (
-  actor: RqgActor,
-  token: TokenDocument | undefined
-): ContextMenu.Item[] => [
+export const spiritMagicMenuOptions = (actor: RqgActor): ContextMenu.Item[] => [
   {
     name: localize("RQG.Game.RollChat"),
     icon: ContextMenuRunes.RollViaChat,
@@ -41,13 +37,7 @@ export const spiritMagicMenuOptions = (
       if (item.data.data.isVariable && item.data.data.points > 1) {
         await item.toChat();
       } else {
-        await SpiritMagicChatHandler.roll(
-          item,
-          item.data.data.points,
-          0,
-          actor,
-          ChatMessage.getSpeaker({ actor: actor, token: token })
-        );
+        await item?.abilityRoll();
       }
     },
   },
