+++ src/actors/item-specific/abstractEmbeddedItem.ts
@@ -1,6 +1,6 @@
 import { RqgItem } from "../../items/rqgItem";
 import { RqgActor } from "../rqgActor";
-import { RqgError } from "../../system/util";
+import { ResultEnum } from "../../data-model/shared/ability";
 
 /**
  * Separate item specific actions that should be done on embedded items when actor _onCreateEmbeddedDocuments etc. is called.
@@ -9,10 +9,17 @@ export abstract class AbstractEmbeddedItem {
   // TODO ***
   // public static init() {}
 
-  static toChat(item: RqgItem): void {}
+  /**
+   * Send this item to a chat message to select option before doing a roll against it.
+   */
+  static async toChat(item: RqgItem): Promise<void> {}
 
-  static async rollFromChat(): Promise<void> {}
-  static async roll(): Promise<void> {}
+  /**
+   * Do a roll immediately with specified options.
+   */
+  static async abilityRoll(item: RqgItem, options: {}): Promise<ResultEnum | undefined> {
+    return;
+  }
 
   static preEmbedItem(actor: RqgActor, item: RqgItem, options: object[], userId: string): void {}
