+++ src/actors/item-specific/rune.ts
@@ -2,9 +2,10 @@ import { AbstractEmbeddedItem } from "./abstractEmbeddedItem";
 import { RqgActor } from "../rqgActor";
 import { RqgItem } from "../../items/rqgItem";
 import { ItemTypeEnum } from "../../data-model/item-data/itemTypes";
-import { activateChatTab, localize } from "../../system/util";
+import { assertItemType, formatModifier, localize } from "../../system/util";
 import { ItemChatFlags } from "../../data-model/shared/rqgDocumentFlags";
 import { ItemChatHandler } from "../../chat/itemChatHandler";
+import { ResultEnum } from "../../data-model/shared/ability";
 
 export class Rune extends AbstractEmbeddedItem {
   // public static init() {
@@ -30,6 +31,23 @@ export class Rune extends AbstractEmbeddedItem {
     await ChatMessage.create(await ItemChatHandler.renderContent(flags));
   }
 
+  static async abilityRoll(
+    runeItem: RqgItem,
+    options: { modifier: number }
+  ): Promise<ResultEnum | undefined> {
+    const chance: number = Number((runeItem?.data.data as any).chance) || 0;
+    let flavor = localize("RQG.Dialog.itemChat.RollFlavor", { name: runeItem.name });
+    if (options.modifier && options.modifier !== 0) {
+      flavor += localize("RQG.Dialog.itemChat.RollFlavorModifier", {
+        modifier: formatModifier(options.modifier),
+      });
+    }
+    const speaker = ChatMessage.getSpeaker({ actor: runeItem.actor ?? undefined });
+    const result = await runeItem._roll(flavor, chance, options.modifier, speaker);
+    await runeItem.checkExperience(result);
+    return result;
+  }
+
   static preUpdateItem(actor: RqgActor, rune: RqgItem, updates: any[], options: any): void {
     if (rune.data.type === ItemTypeEnum.Rune) {
       const chanceResult = updates.find((r) => r["data.chance"] != null);
@@ -38,29 +56,17 @@ export class Rune extends AbstractEmbeddedItem {
       }
       if (rune.data.data.opposingRune) {
         const opposingRune = actor.items.getName(rune.data.data.opposingRune);
-        this.adjustOpposingRuneChance(
-          opposingRune,
-          rune,
-          actor,
-          chanceResult["data.chance"],
-          updates
-        );
+        this.adjustOpposingRuneChance(opposingRune, chanceResult["data.chance"], updates);
       }
     }
   }
 
   private static adjustOpposingRuneChance(
     opposingRune: RqgItem | undefined,
-    item: RqgItem,
-    actor: RqgActor,
     newChance: number,
     updates: object[]
   ) {
-    if (!opposingRune || opposingRune.data.type !== ItemTypeEnum.Rune) {
-      const msg = localize("RQG.Item.Notification.OpposingRuneDoesNotExistWarning");
-      ui.notifications?.warn(msg);
-      return;
-    }
+    assertItemType(opposingRune?.data.type, ItemTypeEnum.Rune);
     const opposingRuneChance = opposingRune.data.data.chance;
     if (newChance + opposingRuneChance !== 100) {
       updates.push({
