+++ src/actors/item-specific/runeMagic.ts
@@ -2,12 +2,14 @@ import { AbstractEmbeddedItem } from "./abstractEmbeddedItem";
 import { ItemTypeEnum } from "../../data-model/item-data/itemTypes";
 import { RqgActor } from "../rqgActor";
 import { RqgItem } from "../../items/rqgItem";
-import { activateChatTab, assertItemType, getGame, localize, RqgError } from "../../system/util";
+import { assertActorType, assertItemType, getGame, localize, RqgError } from "../../system/util";
 import { ItemDataSource } from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/itemData";
 import { systemId } from "../../system/config";
-import { ItemChatFlags, RuneMagicChatFlags } from "../../data-model/shared/rqgDocumentFlags";
-import { ItemChatHandler } from "../../chat/itemChatHandler";
+import { RuneMagicChatFlags } from "../../data-model/shared/rqgDocumentFlags";
 import { RuneMagicChatHandler } from "../../chat/runeMagicChatHandler";
+import { ResultEnum, ResultMessage } from "../../data-model/shared/ability";
+import { ActorTypeEnum } from "../../data-model/actor-data/rqgActorData";
+import { RuneDataPropertiesData } from "../../data-model/item-data/runeData";
 
 export class RuneMagic extends AbstractEmbeddedItem {
   // public static init() {
@@ -18,8 +20,8 @@ export class RuneMagic extends AbstractEmbeddedItem {
   // }
 
   static async toChat(runeMagic: RqgItem): Promise<void> {
-    const eligibleRunes = RuneMagicChatHandler.getEligibleRunes(runeMagic, runeMagic.actor!);
-    const defaultRuneId = RuneMagicChatHandler.getStrongestRune(eligibleRunes)?.id;
+    const eligibleRunes = RuneMagic.getEligibleRunes(runeMagic);
+    const defaultRuneId = RuneMagic.getStrongestRune(eligibleRunes)?.id;
     assertItemType(runeMagic.data.type, ItemTypeEnum.RuneMagic);
     const flags: RuneMagicChatFlags = {
       type: "runeMagicChat",
@@ -40,6 +42,91 @@ export class RuneMagic extends AbstractEmbeddedItem {
     };
 
     await ChatMessage.create(await RuneMagicChatHandler.renderContent(flags));
+  }
+
+  static async abilityRoll(
+    runeMagicItem: RqgItem,
+    options: {
+      runePointsCost: number;
+      magicPointBoost: number;
+      ritualOrMeditation: number;
+      skillAugmentation: number;
+      otherModifiers: number;
+      selectedRuneId?: string;
+    }
+  ): Promise<ResultEnum | undefined> {
+    assertItemType(runeMagicItem.data.type, ItemTypeEnum.RuneMagic);
+    const runeMagicCultId = runeMagicItem?.data.data.cultId;
+    const cult = runeMagicItem.actor?.items.get(runeMagicCultId);
+    assertItemType(cult?.data.type, ItemTypeEnum.Cult);
+    if (!options.selectedRuneId) {
+      const eligibleRunes = RuneMagic.getEligibleRunes(runeMagicItem);
+      options.selectedRuneId = RuneMagic.getStrongestRune(eligibleRunes)?.id ?? "";
+    }
+    const runeItem = runeMagicItem.actor?.items.get(options.selectedRuneId);
+    assertItemType(runeItem?.data.type, ItemTypeEnum.Rune);
+
+    const validationError = RuneMagic.validateData(
+      cult,
+      options.runePointsCost,
+      options.magicPointBoost
+    );
+    if (validationError) {
+      ui.notifications?.warn(validationError);
+      return;
+    }
+
+    const resultMessages: ResultMessage[] = [
+      {
+        result: ResultEnum.Critical,
+        html: localize("RQG.Dialog.runeMagicChat.resultMessageCritical", {
+          magicPointBoost: options.magicPointBoost,
+        }),
+      },
+      {
+        result: ResultEnum.Special,
+        html: localize("RQG.Dialog.runeMagicChat.resultMessageSpecial", {
+          runePointCost: options.runePointsCost,
+          magicPointBoost: options.magicPointBoost,
+        }),
+      },
+      {
+        result: ResultEnum.Success,
+        html: localize("RQG.Dialog.runeMagicChat.resultMessageSuccess", {
+          runePointCost: options.runePointsCost,
+          magicPointBoost: options.magicPointBoost,
+        }),
+      },
+      {
+        result: ResultEnum.Failure,
+        html: localize("RQG.Dialog.runeMagicChat.resultMessageFailure"),
+      },
+      {
+        result: ResultEnum.Fumble,
+        html: localize("RQG.Dialog.runeMagicChat.resultMessageFumble", {
+          runePointCost: options.runePointsCost,
+        }),
+      },
+    ];
+
+    const speaker = ChatMessage.getSpeaker({ actor: runeMagicItem.actor ?? undefined });
+    const result = await runeMagicItem._roll(
+      localize("RQG.Dialog.runeMagicChat.Cast", { spellName: runeMagicItem.name }),
+      Number(runeItem.data.data.chance),
+      options.ritualOrMeditation + options.skillAugmentation + options.otherModifiers,
+      speaker,
+      resultMessages
+    );
+
+    await RuneMagic.handleRollResult(
+      result,
+      options.runePointsCost,
+      options.magicPointBoost,
+      runeItem,
+      runeMagicItem
+    );
+
+    return result;
   }
 
   static onActorPrepareEmbeddedEntities(item: RqgItem): RqgItem {
@@ -167,5 +254,178 @@ export class RuneMagic extends AbstractEmbeddedItem {
       });
       await dialog.render(true);
     });
+  }
+
+  public static validateData(
+    cultItem: RqgItem,
+    runePointsUsed: number,
+    magicPointsBoost: number
+  ): string {
+    assertItemType(cultItem?.data.type, ItemTypeEnum.Cult);
+    if (runePointsUsed > (Number(cultItem.data.data.runePoints.value) || 0)) {
+      return getGame().i18n.format("RQG.Dialog.runeMagicChat.validationNotEnoughRunePoints");
+    } else if (
+      magicPointsBoost > (Number(cultItem.actor?.data.data.attributes?.magicPoints?.value) || 0)
+    ) {
+      return localize("RQG.Dialog.runeMagicChat.validationNotEnoughMagicPoints");
+    } else {
+      return "";
+    }
+  }
+
+  /**
+   * Given a rune spell and an actor, returns the runes that are possible to use for casting that spell.
+   */
+  static getEligibleRunes(runeMagicItem: RqgItem): RqgItem[] {
+    assertItemType(runeMagicItem.data.type, ItemTypeEnum.RuneMagic);
+
+    // The cult from where the spell was learned
+    const cult = runeMagicItem.actor?.items.get(runeMagicItem.data.data.cultId);
+    assertItemType(cult?.data.type, ItemTypeEnum.Cult);
+
+    // Get the name of the "magic" rune.
+    const magicRuneName = getGame().settings.get(systemId, "magicRuneName");
+
+    let usableRuneNames: string[];
+    if (runeMagicItem.data.data.runes.includes(magicRuneName)) {
+      // Actor can use any of the cult's runes to cast
+      // And some cults have the same rune more than once, so de-dupe them
+      usableRuneNames = [...new Set(cult.data.data.runes)];
+    } else {
+      // Actor can use any of the Rune Magic Spell's runes to cast
+      usableRuneNames = [...new Set(runeMagicItem.data.data.runes)];
+    }
+
+    let runesForCasting: RqgItem[] = [];
+    // Get the actor's versions of the runes, which will have their "chance"
+    usableRuneNames.forEach((runeName: string) => {
+      const actorRune = runeMagicItem.actor?.items.getName(runeName);
+      assertItemType(actorRune?.data.type, ItemTypeEnum.Rune);
+      runesForCasting.push(actorRune);
+    });
+
+    return runesForCasting;
+  }
+
+  static getStrongestRune(runeMagicItems: RqgItem[]): RqgItem | undefined {
+    if (runeMagicItems.length === 0) {
+      return undefined;
+    }
+    return runeMagicItems.reduce((strongest: RqgItem, current: RqgItem) => {
+      const strongestRuneChance = (strongest.data.data as RuneDataPropertiesData).chance ?? 0;
+      const currentRuneChance = (current.data.data as RuneDataPropertiesData).chance ?? 0;
+      return strongestRuneChance > currentRuneChance ? strongest : current;
+    });
+  }
+
+  private static async handleRollResult(
+    result: ResultEnum,
+    runePointsUsed: number,
+    magicPointsUsed: number,
+    runeItem: RqgItem,
+    runeMagicItem: RqgItem
+  ): Promise<void> {
+    assertItemType(runeItem.data.type, ItemTypeEnum.Rune);
+    assertItemType(runeMagicItem.data.type, ItemTypeEnum.RuneMagic);
+    const cult = runeMagicItem.actor?.items.get(runeMagicItem.data.data.cultId ?? "");
+    assertItemType(cult?.data.type, ItemTypeEnum.Cult);
+    const isOneUse = runeMagicItem.data.data.isOneUse;
+
+    switch (result) {
+      case ResultEnum.Critical:
+      case ResultEnum.SpecialCritical:
+      case ResultEnum.HyperCritical:
+        // spell takes effect, Rune Points NOT spent, Rune gets xp check, boosting Magic Points spent
+        await RuneMagic.SpendRuneAndMagicPoints(
+          0,
+          magicPointsUsed,
+          runeMagicItem.actor ?? undefined,
+          cult,
+          isOneUse
+        );
+        await runeItem.awardExperience();
+        break;
+
+      case ResultEnum.Success:
+      case ResultEnum.Special:
+        // spell takes effect, Rune Points spent, Rune gets xp check, boosting Magic Points spent
+        await RuneMagic.SpendRuneAndMagicPoints(
+          runePointsUsed,
+          magicPointsUsed,
+          runeMagicItem.actor ?? undefined,
+          cult,
+          isOneUse
+        );
+        await runeItem.awardExperience();
+        break;
+
+      case ResultEnum.Failure:
+        {
+          // spell fails, no Rune Point Loss, if Magic Point boosted, lose 1 Magic Point if boosted
+          const boosted = magicPointsUsed >= 1 ? 1 : 0;
+          await RuneMagic.SpendRuneAndMagicPoints(
+            0,
+            boosted,
+            runeMagicItem.actor ?? undefined,
+            cult,
+            isOneUse
+          );
+        }
+        break;
+
+      case ResultEnum.Fumble:
+        // spell fails, lose Rune Points, if Magic Point boosted, lose 1 Magic Point if boosted
+        const boosted = magicPointsUsed >= 1 ? 1 : 0;
+        await RuneMagic.SpendRuneAndMagicPoints(
+          runePointsUsed,
+          boosted,
+          runeMagicItem.actor ?? undefined,
+          cult,
+          isOneUse
+        );
+        break;
+
+      default:
+        const msg = "Got unexpected result from roll in runeMagicChat";
+        ui.notifications?.error(msg);
+        throw new RqgError(msg);
+    }
+  }
+
+  private static async SpendRuneAndMagicPoints(
+    runePoints: number,
+    magicPoints: number,
+    actor: RqgActor | undefined,
+    cult: RqgItem,
+    isOneUse: boolean
+  ) {
+    assertItemType(cult.data.type, ItemTypeEnum.Cult);
+    assertActorType(actor?.data.type, ActorTypeEnum.Character);
+    // At this point if the current Rune Points or Magic Points are zero
+    // it's too late. That validation happened earlier.
+    const newRunePointTotal = (cult.data.data.runePoints.value || 0) - runePoints;
+    const newMagicPointTotal = (actor?.data.data.attributes.magicPoints.value || 0) - magicPoints;
+    let newRunePointMaxTotal = cult.data.data.runePoints.max || 0;
+    if (isOneUse) {
+      newRunePointMaxTotal -= runePoints;
+      if (newRunePointMaxTotal < (cult.data.data.runePoints.max || 0)) {
+        ui.notifications?.info(
+          localize("RQG.Dialog.runeMagicChat.SpentOneUseRunePoints", {
+            actorName: actor?.name,
+            runePoints: runePoints,
+            cultName: cult.name,
+          })
+        );
+      }
+    }
+    const updateCultItemRunePoints: DeepPartial<ItemDataSource> = {
+      _id: cult?.id,
+      data: { runePoints: { value: newRunePointTotal, max: newRunePointMaxTotal } },
+    };
+    await actor?.updateEmbeddedDocuments("Item", [updateCultItemRunePoints]);
+    const updateActorMagicPoints = {
+      data: { attributes: { magicPoints: { value: newMagicPointTotal } } },
+    };
+    await actor?.update(updateActorMagicPoints);
   }
 }
