+++ src/chat/itemChatHandler.ts
@@ -1,6 +1,4 @@
-import { Ability, ResultEnum } from "../data-model/shared/ability";
 import {
-  formatModifier,
   getGame,
   localize,
   RqgError,
@@ -14,7 +12,6 @@ import {
   getRequiredRqgActorFromUuid,
 } from "../system/util";
 import { RqgActor } from "../actors/rqgActor";
-import { ChatSpeakerDataProperties } from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/chatSpeakerData";
 import { RqgChatMessageFlags } from "../data-model/shared/rqgDocumentFlags";
 import { RqgItem } from "../items/rqgItem";
 import { ChatMessageDataConstructorData } from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/data.mjs/chatMessageData";
@@ -28,41 +25,11 @@ export class ItemChatHandler {
   public static async rollFromChat(chatMessage: RqgChatMessage): Promise<void> {
     const flags = chatMessage.data.flags.rqg;
     assertChatMessageFlagType(flags?.type, "itemChat");
-    const actor = await getRequiredRqgActorFromUuid<RqgActor>(flags.chat.actorUuid);
-    const token = await getDocumentFromUuid<TokenDocument>(flags.chat.tokenUuid);
-    const speaker = ChatMessage.getSpeaker({ actor: actor, token: token });
     const item = (await getRequiredDocumentFromUuid(flags.chat.itemUuid)) as RqgItem | undefined;
     requireValue(item, "Couldn't find item on item chat message");
     const { modifier } = await ItemChatHandler.getFormDataFromFlags(flags);
 
-    await ItemChatHandler.roll(item, modifier, actor, speaker);
-  }
-
-  public static async roll(
-    item: RqgItem,
-    modifier: number,
-    actor: RqgActor,
-    speaker: ChatSpeakerDataProperties
-  ): Promise<void> {
-    const chance: number = Number((item?.data.data as any).chance) || 0;
-    let flavor = localize("RQG.Dialog.itemChat.RollFlavor", { name: item.name });
-    if (modifier !== 0) {
-      flavor += localize("RQG.Dialog.itemChat.RollFlavorModifier", {
-        modifier: formatModifier(modifier),
-      });
-    }
-    const result = await Ability.roll(flavor, chance, modifier, speaker);
-    await ItemChatHandler.checkExperience(actor, item, result);
-  }
-
-  public static async checkExperience(
-    actor: RqgActor,
-    item: RqgItem,
-    result: ResultEnum
-  ): Promise<void> {
-    if (result <= ResultEnum.Success && !(item.data.data as any).hasExperience) {
-      await actor.AwardExperience(item.id);
-    }
+    await item.abilityRoll({ modifier: modifier });
   }
 
   public static async renderContent(
